+++ cmd/frontend/internal/app/jscontext/jscontext.go
@@ -64,8 +64,8 @@ type JSContext struct {
 
 	SourcegraphDotComMode bool `json:"sourcegraphDotComMode"`
 
-	DiscussionsEnabled  bool `json:"discussionsEnabled"`
-	AccessTokensEnabled bool `json:"accessTokensEnabled"`
+	DiscussionsEnabled bool                `json:"discussionsEnabled"`
+	AccessTokensAllow  conf.AccessTokAllow `json:"accessTokensAllow"`
 
 	AllowSignup bool `json:"allowSignup"`
 
@@ -151,8 +151,8 @@ func NewJSContextFromRequest(req *http.Request) JSContext {
 
 		// Experiments. We pass these through explicitly so we can
 		// do the default behavior only in Go land.
-		DiscussionsEnabled:  conf.DiscussionsEnabled(),
-		AccessTokensEnabled: conf.AccessTokensEnabled(),
+		DiscussionsEnabled: conf.DiscussionsEnabled(),
+		AccessTokensAllow:  conf.AccessTokensAllow(),
 
 		ResetPasswordEnabled: userpasswd.ResetPasswordEnabled(),