+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 - gitserver periodically removes stale lock files that git can leave behind.
 - Commits with empty trees no longer return 404.
 - Clients (browser/editor extensions) can now query configuration details from the `ClientConfiguration` GraphQL API.
+- The config field `auth.accessTokens.allow` allows or restricts use of access tokens. It can be set to one of three values: "all-users-create" (the default), "none" (all access tokens are disabled), and "site-admin-create" (access tokens are enabled, but only site admins can create new access tokens). The field `auth.disableAccessTokens` is now deprecated in favor of this new field.
 
 ### Fixed